+++ api/api_test.go
@@ -15,6 +15,7 @@ import (
 	"time"
 
 	"github.com/hashicorp/consul/sdk/testutil"
+	"github.com/hashicorp/consul/sdk/testutil/retry"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -52,11 +53,20 @@ func makeClientWithConfig(
 	if cb1 != nil {
 		cb1(conf)
 	}
+
 	// Create server
-	server, err := testutil.NewTestServerConfigT(t, cb2)
-	if err != nil {
-		t.Fatal(err)
+	var server *testutil.TestServer
+	var err error
+	retry.RunWith(retry.ThreeTimes(), t, func(r *retry.R) {
+		server, err = testutil.NewTestServerConfigT(t, cb2)
+		if err != nil {
+			r.Fatal(err)
+		}
+	})
+	if server.Config.Bootstrap {
+		server.WaitForLeader(t)
 	}
+
 	conf.Address = server.HTTPAddr
 
 	// Create client